style: 리스트 여백 수정
